YH-148: refactor header
