+++ src/entities/authentication/selectors/authSelectors.ts
@@ -1,0 +1,3 @@
+import { State } from '@/shared/config/store/State';
+
+export const getAuthData = (state: State) => state.auth.user;
