+++ src/features/authentication/logout/Logout/index.ts
@@ -1,0 +1,1 @@
+export { Logout } from './ui/Logout';
