+++ src/features/authentication/logout/Logout/ui/Logout.module.css
@@ -1,0 +1,23 @@
+.button-wrapper {
+    display: flex;
+    align-items: center;
+    padding: 6px;
+    border-radius: 12px;
+}
+
+.button-wrapper:hover {
+    background-color: var(--color-background-gray-2);
+    color: var(--color-purple-800);
+}
+
+.button {
+    display: block;
+    width: 100%;
+    cursor: pointer;
+    color: var(--color-black-text);
+}
+
+.icon {
+    margin-right: 15px;
+    max-width: 24px;
+}
