--- src/features/header/UserProfile/index.ts
@@ -1,1 +1,0 @@
-export { UserProfile } from './ui/UserProfile/UserProfile';
