+++ src/shared/ui/AuthAvatarFrame/AuthAvatarFrame.module.css
