+++ src/shared/ui/AuthAvatarFrame/index.ts
@@ -1,0 +1,1 @@
+export { AuthAvatarFrame } from './AuthAvatarFrame';
