+++ src/widgets/FeedbackBlock/ui/FeedbackBlock.tsx
@@ -2,7 +2,7 @@ import { useRef, MouseEvent, useCallback, useEffect } from 'react';
 
 import { mockFeedback } from '@/shared/const/mockFeedback';
 import { useObserver } from '@/shared/libs/hooks/useObserver';
-import { makeCarouselInfinite } from '@/shared/libs/makeCarouselInfinite';
+import { makeCarouselInfinite } from '@/shared/libs/utils/makeCarouselInfinite';
 import { Button } from '@/shared/ui/Button';
 import { FeedbackCard } from '@/shared/ui/FeedbackCard';
