+++ src/widgets/Header/index.ts
@@ -1,1 +1,1 @@
-export { Header } from './ui/Header';
+export { Header } from './ui/Header/Header';
