+++ src/widgets/Header/ui/AuthorizedBlock/AuthorizedBlock.tsx
@@ -3,41 +3,30 @@
 import { useState } from 'react';
 import { Button } from 'yeahub-ui-kit';
 
-import LogoutIcon from '@/shared/assets/icons/logout-icon.svg';
 import ProfileIcon from '@/shared/assets/icons/profile-icon.svg';
-import { useAppDispatch } from '@/shared/hooks/useAppDispatch/useAppDispatch';
+import { AuthAvatarFrame } from '@/shared/ui/AuthAvatarFrame';
 
-import { useLazyLogoutQuery, authActions } from '@/entities/authentication';
+import { Logout } from '@/features/authentication/logout/Logout';
 
-import { Avatar } from '../Avatar/Avatar';
-
-import styles from './UserProfile.module.css';
+import styles from './AuthorizedBlock.module.css';
 
 interface UserProfileProps {
 	firstName: string;
 	avatarURL: string | null;
 }
 
-export const UserProfile = ({ firstName, avatarURL }: UserProfileProps) => {
+export const AuthorizedBlock = ({ firstName, avatarURL }: UserProfileProps) => {
 	const [isOpen, setIsOpen] = useState(false);
-	const dispatch = useAppDispatch();
-	const [trigger] = useLazyLogoutQuery();
 
 	const handleDropdownClick = () => {
-		console.log(isOpen);
 		setIsOpen(!isOpen);
-	};
-
-	const onLogout = () => {
-		trigger();
-		dispatch(authActions.logOut());
 	};
 
 	return (
 		<div className={styles.wrapper}>
 			<div role="banner" className={styles['user-wrapper']} onClick={handleDropdownClick}>
 				<p className={styles['user-name']}>{firstName}</p>
-				<Avatar link={avatarURL || ''} />
+				<AuthAvatarFrame link={avatarURL || ''} />
 			</div>
 			{isOpen && (
 				<div className={styles.dropdown}>
@@ -47,12 +36,7 @@ export const UserProfile = ({ firstName, avatarURL }: UserProfileProps) => {
 							Мой профиль
 						</Button>
 					</div>
-					<div className={styles['button-wrapper']}>
-						<LogoutIcon className={styles.icon} />
-						<Button tagName="a" theme="link" className={styles.button} onClick={onLogout}>
-							Выйти
-						</Button>
-					</div>
+					<Logout />
 				</div>
 			)}
 		</div>
